// src/team-requests/team-requests.type.ts
export enum ETeamRequestTypes {
    JOIN_THE_TEAM = 'Join the team',
    LEAVE_THE_TEAM = 'Leave the team',
    MOVE_TO_ANOTHER_TEAM = 'Move to another team',
    MANAGER_POST = 'Manager post'
}

export enum ETeamRequestStatusType{
    AWAITING = 'Awaiting',
    ACCEPTED = 'Accepted',
    DECLINED = 'Declined'
}